+++ integrations/consensus/src/main/java/org/pragmatica/consensus/net/netty/NettyClusterNetwork.java
@@ -30,8 +30,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
@@ -68,7 +66,6 @@ public class NettyClusterNetwork implements ClusterNetwork {
     private final Supplier<List<ChannelHandler>> handlers;
     private final MessageRouter router;
     private final AtomicReference<Server> server = new AtomicReference<>();
-    private final Executor executor;
 
     enum ViewChangeOperation {
         ADD,
@@ -80,11 +77,6 @@ public class NettyClusterNetwork implements ClusterNetwork {
                                Serializer serializer,
                                Deserializer deserializer,
                                MessageRouter router) {
-        var processors = Runtime.getRuntime()
-                                .availableProcessors();
-        this.executor = Executors.newFixedThreadPool(processors > 1
-                                                     ? processors
-                                                     : 2);
         this.self = topologyManager.self();
         this.topologyManager = topologyManager;
         this.router = router;
@@ -322,7 +314,7 @@ public class NettyClusterNetwork implements ClusterNetwork {
 
     @Override
     public <M extends ProtocolMessage> void broadcast(M message) {
-        peerLinks.forEach((peerId, channel) -> executor.execute(() -> sendToChannel(peerId, message, channel)));
+        peerLinks.forEach((peerId, channel) -> sendToChannel(peerId, message, channel));
     }
 
     private void processViewChange(ViewChangeOperation operation, NodeId peerId) {